petitboot-powerpc: persist kexec_method option via nvram
When the kernel is running in a secure mode, a regular kexec_load should
be disabled. To avoid having to tick the box in the system config menu
for every single boot, this patch introduces an nvram field for storing
the state of the "Use file_load" config option across reboots.

Signed-off-by: Eric Richter <erichte@linux.vnet.ibm.com>